requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.0.2
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
email-validator==2.1.1
PyJWT==2.8.0
httpx==0.24.1
pytest==7.4.3
pytest-asyncio==0.21.1
google-generativeai>=0.3.0
beautifulsoup4>=4.12.0
aiohttp>=3.9.0
lxml>=4.9.0